+++ packages/@tinacms/cli/src/next/config-manager.ts
@@ -8,7 +8,10 @@ import * as dotenv from 'dotenv'
 import normalizePath from 'normalize-path'
 import chalk from 'chalk'
 import { logger } from '../logger'
-import { loadViteConfig, resolveDynamicAliases } from '../utils/alias-helpers'
+import {
+  loadViteConfig,
+  resolveDynamicAliases,
+} from '../utils/resolve-alias-helpers'
 import { aliasPath } from '../utils/esbuild-alias-lib/esbuild-plugin-alias-path'
 
 export const TINA_FOLDER = 'tina'
